+++ sphinx/source/changelog.rst
@@ -3,6 +3,8 @@ Changelog
 
 8.9b1
 -----
+* Added support to the ``--sim-load`` option of :ref:`tap2sna.py` for TZX loops
+  (block types 0x24 and 0x25)
 * Fixed how the ``--sim-load`` option of :ref:`tap2sna.py` transitions from a
   tape block that ends with data to the next block both when there is a
   specified pause between them and when there is no pause between them